build: Split Makefile build system into multiple files
This commit makes a new build/ folder which contains auxiliary
Makefiles that are used to assist the build.

In theory, later on, (hinted by the documentation of build/core.mk),
it may be possible to even add IOP target support, but for now, I have
not yet added this to the buildsystem. It should be relatively easy, however.