build: Output products into object and binary directories
Objects and binaries are now put in neat bin/ and obj/ directories matching the build configuration. This also means that a release and debug build are allowed to coexist, since they won't overwrite files related to each other.

This also thins out the gitignore, and removes build/ from it since it is now a valid directory (where we put the Make buildsystem files).