build: Add `objonly`/`objflush` targets
This allows tooling (or developers) to build a single object file simply by specifying the object to compile and running the "objonly" target.

For example: `make COMPILEME=bis.o objonly` will compile only bis.o and also output the path to the final object file, useful for tools which want to scan for it.

The "objflush" target also exists for if a clean build of the object file is required.