chore(build): Tell driver where compiler executables are explicitly
The -V option is "used to control which version of gcc to run" (according to GCC2 documentation; this parameter doesn't exist in later GCC driver versions), but it adds the selected version to the usual /usr/local/sce/ee/gcc/lib... path, which doesn't really exist; thus we get warnings telling us that it was never used.

Instead, the -B option does what we want, allowing us to directly override the path to driver executables, without any spurious warnings.